### 2024-05-25 #6783 @editorialbot
[PRE REVIEW]: sourmash v4: a multitool to quickly search, compare, and analyze genomic and metagenomic data sets
**Submitting author:** @bluegenes (<a href="http://orcid.org/0000-0002-2942-5331">N. Tessa Pierce-Ward</a>)
**Repository:** https://github.com/sourmash-bio/sourmash
**Branch with paper.md** (empty if default branch):
**Version:** v4.8.8
