### 2024-05-31 #6783 @majensen
Hi @bede - would you be willing to review [this work](#6783) for [JOSS](https://joss.theoj.org)? Thanks for your consideration -- @majensen
