### 2024-06-05 #6830 @lwasser
hey @majensen  👋🏻 just to add to  @bluegenes comment above, this package has already been reviewed by pyOpenSci. as such you will only need to review the paper / scope etc. The code should not be reviewed again. I am adding this note because i see two reviewers have been assigned. in the past, credit has been given to the reviewers on the pyOpenSci side of things who in this case are.

@LilyAnderssonLee
@elais

@arfon can also answer any questions if you have them! 🙌🏻
