### 2024-06-06 #6830 @lwasser
@majensen 😆 i believe it is. i'm just curious - from your perspective as an editor - what could we (pyopensci / joss) do differently that would help make it more clear that is it a pyopensci fast track when such a package enters into your review process? and also - thank you!!
