### 2024-06-06 #6830 @majensen
@lwasser well unless I missed it in the autogenerated material (very possible), I think it was @bluegenes herself who informed me in the thread. I wonder @openjournals/dev if there is a natural way to interconnect between the journals (one of the paper metadata files?), so that maybe a tag is autoadded or some other info is provided to give the editor a heads-up. I admit this my first time handling one of these over almost 5yr.
