### 2024-06-06 #6830 @lwasser
ahhhh good to know. and so good that @bluegenes posted here (and i was tagged in our slack as well!) . an auto label would be really nice and would make things easier. a check box in the review submission that triggers the bot to create a pyOpenSci approved label? that is a great idea.
