### 2024-06-10 #6830 @majensen
Hi @bluegenes  - you are a generous and warm hearted primary author - for which you will be penalized slightly. Can the Zenodo archive be updated with the full author list, and its title be brought in line with the title of the submitted paper? This is a JOSS editorial preference (which I can fight with with associate editor in chief about if you like).
