### 2024-06-10 #6830 @majensen
I would also capitalize the article after the colon in the title: "sourmash v4: A multitool...".
