### 2024-06-10 #6830 @bluegenes
> Hi @bluegenes - you are a generous and warm hearted primary author - for which you will be penalized slightly. Can the Zenodo archive be updated with the full author list, and its title be brought in line with the title of the submitted paper? This is a JOSS editorial preference (which I can fight with with associate editor in chief about if you like).

Thank you, we wanted to make sure everyone gets credit for their contributions!

Looks like we could update the authorship by including a `.zenodo.json` file (much simpler given the number of authors we have!), but then I think this would be updated as part of a new release. We will need that anyway to update the title -- would that work? cc @ctb
