### 2024-06-10 #6830 @ctb
OK, we verified that the `.zenodo.json` works to set the author list - see [sourmash v4.8.9rc2 on zenodo](https://zenodo.org/records/11556174). And the paper title was fixed in https://github.com/sourmash-bio/sourmash/pull/3203. So if we release sourmash v4.8.9 then it will be updated properly.
