### 2024-06-11 #6830 @majensen
Great - can we now entitle the Zenodo archive "sourmash v4: A multitool to quickly search, compare, and analyze genomic and metagenomic data sets"? Again it's a JOSS policy, but we can argue for a exception if you like.
