### 2024-06-13 #6830 @majensen
Ok team - can you consider the following minor updates (line numbers in PDF in previous post):
l. 36 - (?) where ref should be
l. 41 - “flexible set of programmatic functionality” -> “flexible set of programmatic tools” [?]
l. 47 - PIG-PARADIGM gets gold medal for the farthest backed backronym.
l. 51. - “with the DHS” -> “with the US Department of Homeland Security (DHS)” (I assume)
l. 51. - “The US Government” -> “The US Government (USG)…”
l. 51 - how about a comma after each “retains” in this sentence. I know it’s boilerplate.
l. 65 - Why is this title italicized and not the others? (probable answer: there is no journal and bibtex thinks the title is the journal)
