### 2024-06-20 #6830 @bluegenes
Hi @majensen,

Thanks for your review!

> l. 47 - PIG-PARADIGM gets gold medal for the farthest backed backronym. (or whatever)

agreed! :joy:

I've fixed/updated the following:
> l. 36 - (?) where ref should be
> l. 41 - “flexible set of programmatic functionality” -> “flexible set of programmatic tools” [?]
> l. 65 - Why is this title italicized and not the others? (probable answer: there is no journal and bibtex thinks the title is the journal)

However, line 51 was given to us verbatim by our DHS coauthor, and I believe the manuscript would need to go through another DHS review if we change it. Can we keep as is?

> l. 51. - “with the DHS” -> “with the US Department of Homeland Security (DHS)” (I assume)
> l. 51. - “The US Government” -> “The US Government (USG)…”
> l. 51 - how about a comma after each “retains” in this sentence. I know it’s boilerplate.

Shall we do another release that includes these updates?
