### 2024-06-21 #6830 @ctb
> @bluegenes can you real quick look at the suggestion by editorialbot for the DOI ref above?

That is the correct reference, and it renders correctly in the PDF. Is there something we should do about it? (It's a valid DOI, it resolves correctly, etc - not sure why it's listed as MISSING?)

thanks!!
