### 2024-06-21 #6830 @ctb
Proof looks good to me!
