### 2024-06-23 #6783 @Kevin-Mattheus-Moerman
Apologies @bede @amoeba @majensen because this has been reviewed already here: https://github.com/pyOpenSci/software-submission/issues/129 we follow a different procedure. Thanks anyway for being willing to help!
