### 2024-06-23 #6783 @Kevin-Mattheus-Moerman
This was reviewed here; https://github.com/pyOpenSci/software-submission/issues/129
