### 2024-06-23 #6830 @Kevin-Mattheus-Moerman
@bluegenes apologies for the kinks in the cable with this review as I initially missed this was a pyOpenSci related submission. All is good/cleared up now and we are good to proceed. As AEiC for JOSS I will now help to process this submission for acceptance in JOSS. I have checked the pyOpenSci review, your repository, the archive link, and the paper. Most seems in order, however the below are some points that require your attention:

On the archive:
- [ ] I updated the archive as the one listed was for the wrong version. Please can you confirm this is correct, i.e. this review is for tagged release version [`v4.8.9`](https://github.com/sourmash-bio/sourmash/releases/tag/v4.8.9) which is archived here [10.5281/zenodo.11557883](https://zenodo.org/records/11557883).
- [ ] Please edit the ZENODO listed license to match your software license. It currently lists this:
![image](https://github.com/openjournals/joss-reviews/assets/8392709/760f44e4-adfc-43c1-b38e-0b1fb8871772)
But it should refer to you actual BSD-3-Clause license. You need to process this change by manually editing your archive entry.

On the paper:
- [ ] Please add city, state, country to all affiliations, and please spell out country names, so use United States of America in full.
