### 2024-06-23 #6830 @ctb
I changed the LICENSE just now - good catch, we'll have to edit the .zenodo.json to fix that more permanently!

In re the authors, we do not have that level of affiliation for many of them, and would need to put in a lot of work to get it - with no guarantee of success, since this is now a 9 year old package. We do have their ORCIDs, of course, but not all (most?) of the ORCIDs have public affiliation information. Could you confirm this is necessary for publication?

Last but not least - if we do need to update the information on that scale, we will probably do so via .zenodo.json and then would release a new version, v4.8.10. So that would be the version to list then.

thanks!
