### 2024-06-25 #6830 @ctb
> Since this is only an affiliation rendering update as a minor change in `paper.md` there is no need to mint a new version etc. Hope this helps. Hope you can make this minor change where possible.

thank you so much! that makes a lot of sense :). I'm traveling and it may take us a day or three to make these changes, but will do!
