### 2024-06-28 #6830 @Kevin-Mattheus-Moerman
@ctb thanks for making those changes. I have checked the review over at https://github.com/pyOpenSci/software-submission/issues/129, and I have also checked your repository, the archive link, and the paper. All is now in order, so I will now proceed to accept this submission.
